refactor(catalog): lift create/delete/restore helpers into entity_crud.py
Pulls the create / delete / restore / delete-preview wiring out of
taxonomy.py into a new apps/catalog/api/entity_crud.py module. Entity
shape is now injected as (detail_qs, serialize_detail, response_schema)
so the same helpers can serve richer entities (Theme, GameplayFeature,
Series, Franchise, System) without duplicating the rate-limit /
validation / blocker / restore plumbing.

Zero behavior change — all 8 shipped taxonomy entities re-route through
the generic helpers via local adapters that preserve the existing
(TaxonomySchema, _serialize_taxonomy, _taxonomy_detail_qs) triple.
Public schema names (TaxonomyCreateSchema, TaxonomyDeleteSchema,
TaxonomyRestoreSchema, TaxonomyDeletePreviewSchema,
TaxonomyDeleteResponseSchema) are intentionally preserved so the
OpenAPI surface and the shipped frontend delete loaders don't churn.

Test coverage: existing apps/catalog/tests/test_api_taxonomy_crud.py
passes unchanged (the regression guard for the refactor).
